ready the client for sharing with options
