+++ Intrinio.Realtime/ClientStats.cs
@@ -1,4 +1,4 @@
-namespace Intrinio.Realtime.Equities;
+namespace Intrinio.Realtime;
 
 using System;
 
@@ -8,22 +8,18 @@ public class ClientStats
     private readonly UInt64 _socketTextMessages;
     private readonly int _queueDepth;
     private readonly UInt64 _eventCount;
-    private readonly UInt64 _tradeCount;
-    private readonly UInt64 _quoteCount;
     private readonly int _queueCapacity;
     private readonly int _overflowQueueDepth;
     private readonly int _overflowQueueCapacity;
     private readonly int _droppedCount;
     private readonly int _overflowCount;
 
-    public ClientStats(UInt64 socketDataMessages, UInt64 socketTextMessages, int queueDepth, UInt64 eventCount, UInt64 tradeCount, UInt64 quoteCount, int queueCapacity, int overflowQueueDepth, int overflowQueueCapacity, int droppedCount, int overflowCount)
+    public ClientStats(UInt64 socketDataMessages, UInt64 socketTextMessages, int queueDepth, UInt64 eventCount, int queueCapacity, int overflowQueueDepth, int overflowQueueCapacity, int droppedCount, int overflowCount)
     {
         _socketDataMessages = socketDataMessages;
         _socketTextMessages = socketTextMessages;
         _queueDepth = queueDepth;
         _eventCount = eventCount;
-        _tradeCount = tradeCount;
-        _quoteCount = quoteCount;
         _queueCapacity = queueCapacity;
         _overflowQueueDepth = overflowQueueDepth;
         _overflowQueueCapacity = overflowQueueCapacity;
@@ -64,16 +60,6 @@ public class ClientStats
     public UInt64 EventCount()
     {
         return _eventCount;
-    }
-    
-    public UInt64 TradeCount()
-    {
-        return _tradeCount;
-    }
-    
-    public UInt64 QuoteCount()
-    {
-        return _quoteCount;
     }
 
     public int DroppedCount()
