+++ Intrinio.Realtime/Equities/IEquitiesWebSocketClient.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Threading.Tasks;
 
 namespace Intrinio.Realtime.Equities;
 
@@ -6,13 +7,16 @@ public interface IEquitiesWebSocketClient
 {
     public Action<Trade> OnTrade { set; }
     public Action<Quote> OnQuote { set; }
-    public void Join();
-    public void Join(string channel, bool? tradesOnly);
-    public void Join(string[] channels, bool? tradesOnly);
-    public void Leave();
-    public void Leave(string channel);
-    public void Leave(string[] channels);
-    public void Stop();
+    public Task Join();
+    public Task Join(string channel, bool? tradesOnly);
+    public Task Join(string[] channels, bool? tradesOnly);
+    public Task Leave();
+    public Task Leave(string channel);
+    public Task Leave(string[] channels);
+    public Task Stop();
+    public Task Start();
     public ClientStats GetStats();
-    public void Log(string message, params object[] parameters);
+    public UInt64 TradeCount { get; }
+    public UInt64 QuoteCount { get; }
+    public void LogMessage(LogLevel logLevel, string messageTemplate, params object[] propertyValues);
 }
